Fix restoration of "defflow" setting on attachment.
Before this, the value of "defflow" would be set to "auto" on
attachment, no matter what its value was before. Already opened windows
would have the correct flow settings (i.e., "on", "off", or "auto"), but
newly created windows in a reattached session would always have their
flow set to "auto". This manifested, for example, with "screen -x" and
the lockscreen command.